# prototipo/openapi.yaml
openapi: 3.0.3

info:
  title: API Pública de Coleta de Resíduos Urbanos
  description: >
    Contrato de interface (IDL) para consulta pública do serviço de coleta de
    resíduos urbanos do município de Feira de Santana - BA.
    Permite consultar o cronograma de coleta por bairro, localizar ecopontos
    e registrar ocorrências (acúmulo irregular de resíduos, coleta não realizada).
  version: 1.0.0
  contact:
    name: Squad Sistemas Distribuídos 2026.2 - UNIFAN

servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento

tags:
  - name: Bairros
    description: Consulta de bairros e cronogramas de coleta
  - name: Ecopontos
    description: Pontos de entrega voluntária de recicláveis
  - name: Ocorrências
    description: Registro e acompanhamento de reclamações do cidadão

paths:
  /bairros:
    get:
      tags: [Bairros]
      summary: Lista todos os bairros atendidos pelo serviço de coleta
      parameters:
        - name: zona
          in: query
          required: false
          description: Filtra os bairros por zona da cidade
          schema:
            type: string
            enum: [Central, Norte, Sul, Leste, Oeste]
      responses:
        '200':
          description: Lista de bairros retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 6
                  bairros:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bairro'

  /bairros/{id_bairro}:
    get:
      tags: [Bairros]
      summary: Retorna os dados cadastrais de um bairro específico
      parameters:
        - $ref: '#/components/parameters/IdBairro'
      responses:
        '200':
          description: Bairro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bairro'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

  /bairros/{id_bairro}/coleta:
    get:
      tags: [Bairros]
      summary: Retorna o cronograma e a próxima coleta prevista para o bairro
      description: >
        Recurso principal da API. Calcula, a partir da data atual do servidor,
        qual será a próxima coleta domiciliar e a próxima coleta seletiva.
      parameters:
        - $ref: '#/components/parameters/IdBairro'
      responses:
        '200':
          description: Cronograma calculado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CronogramaColeta'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

  /bairros/{id_bairro}/status:
    get:
      tags: [Bairros]
      summary: Retorna o status da coleta do dia corrente no bairro
      parameters:
        - $ref: '#/components/parameters/IdBairro'
      responses:
        '200':
          description: Status retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusColeta'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

  /ecopontos:
    get:
      tags: [Ecopontos]
      summary: Lista os pontos de entrega voluntária de recicláveis
      parameters:
        - name: material
          in: query
          required: false
          description: Filtra ecopontos que recebem determinado material
          schema:
            type: string
            enum: [vidro, papel, plastico, metal, eletronico, oleo, entulho]
      responses:
        '200':
          description: Lista de ecopontos retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  ecopontos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ecoponto'

  /ocorrencias:
    post:
      tags: [Ocorrências]
      summary: Registra uma ocorrência sobre o serviço de coleta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OcorrenciaRequest'
      responses:
        '201':
          description: Ocorrência registrada e protocolo gerado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ocorrencia'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'

  /ocorrencias/{protocolo}:
    get:
      tags: [Ocorrências]
      summary: Consulta uma ocorrência pelo número de protocolo
      parameters:
        - name: protocolo
          in: path
          required: true
          schema:
            type: string
          example: OC-2026-0001
      responses:
        '200':
          description: Ocorrência encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ocorrencia'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

components:
  parameters:
    IdBairro:
      name: id_bairro
      in: path
      required: true
      description: Identificador único do bairro
      schema:
        type: string
      example: BR001

  responses:
    NaoEncontrado:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    RequisicaoInvalida:
      description: Dados enviados são inválidos ou incompletos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'

  schemas:
    Bairro:
      type: object
      properties:
        id_bairro:
          type: string
          example: BR001
        nome:
          type: string
          example: Centro
        zona:
          type: string
          example: Central
        populacao_estimada:
          type: integer
          example: 18400
        empresa_responsavel:
          type: string
          example: Sustentare Saneamento

    CronogramaColeta:
      type: object
      properties:
        id_bairro:
          type: string
        bairro:
          type: string
        consultado_em:
          type: string
          format: date-time
        coleta_domiciliar:
          $ref: '#/components/schemas/DetalheColeta'
        coleta_seletiva:
          $ref: '#/components/schemas/DetalheColeta'

    DetalheColeta:
      type: object
      properties:
        dias_semana:
          type: array
          items:
            type: string
          example: [segunda, quarta, sexta]
        turno:
          type: string
          enum: [matutino, vespertino, noturno]
        proxima_coleta:
          type: string
          format: date
          example: '2026-09-21'
        dias_restantes:
          type: integer
          example: 3

    StatusColeta:
      type: object
      properties:
        id_bairro:
          type: string
        bairro:
          type: string
        data:
          type: string
          format: date
        ha_coleta_hoje:
          type: boolean
        tipo:
          type: string
          enum: [domiciliar, seletiva, nenhuma]
        status_operacional:
          type: string
          enum: [AGENDADA, EM_ROTA, CONCLUIDA, SEM_COLETA]
        mensagem:
          type: string

    Ecoponto:
      type: object
      properties:
        id_ecoponto:
          type: string
        nome:
          type: string
        endereco:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        materiais_aceitos:
          type: array
          items:
            type: string
        horario_funcionamento:
          type: string

    OcorrenciaRequest:
      type: object
      required: [id_bairro, tipo, descricao]
      properties:
        id_bairro:
          type: string
          example: BR003
        tipo:
          type: string
          enum: [coleta_nao_realizada, acumulo_irregular, container_danificado, descarte_entulho]
        descricao:
          type: string
          example: Resíduos acumulados na esquina da Rua Cel. Silva há três dias.
        solicitante:
          type: string
          example: Lucas

    Ocorrencia:
      type: object
      properties:
        protocolo:
          type: string
          example: OC-2026-0001
        id_bairro:
          type: string
        bairro:
          type: string
        tipo:
          type: string
        descricao:
          type: string
        solicitante:
          type: string
        status:
          type: string
          enum: [ABERTA, EM_ANALISE, RESOLVIDA]
        registrada_em:
          type: string
          format: date-time

    Erro:
      type: object
      properties:
        erro:
          type: string
        mensagem:
          type: string